% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNumbers.R
\name{cleanNumbers}
\alias{cleanNumbers}
\alias{cleanLessthan}
\alias{cleanGreaterthan}
\alias{cleanK}
\alias{cleanStar}
\alias{cleanPlus}
\alias{cleanCruft}
\title{Strip all non-numeric characters from string formated test scores.}
\usage{
cleanNumbers(c)

cleanLessthan(c)

cleanGreaterthan(c)

cleanK(c)

cleanStar(c)

cleanPlus(c)

cleanCruft(c)
}
\arguments{
\item{c}{A character vector of potentially numeric values.}
}
\value{
A character vector
}
\description{
Remove non-numeric characters from otherwise numeric standardized
test scores.
}
\details{
Remove non-numeric characters that sometimes occur in otherwise numeric standardized
test scores. Several characters are handled specially. These four characters "<>+*"
are simply removed. The letter "K" is \emph{\strong{replaced}} with "0" (zero). All other non-numeric
characters are removed. A count of each type of character removed or replaced is
printed to the console.

Advantages over \code{\link{readr::parse_number()}} are that it provides
explicit alerts as to what characters are being replaced and in what quantity. Useful
cases where you are working with unfamiliar data (e.g., client data, or found data).

\code{cleanLessthan()} removes "<" from ordinal values at end of range in anticipation of converting to numeric.
These sometimes occur in otherwise numeric standardized test scores.
Report the number of such characters removed.

\code{cleanGreaterthan()} removes ">" from ordinal values at end of range in anticipation of converting to numeric.
These sometimes occur in otherwise numeric standardized test scores.
Report the number of such characters removed.

\code{cleanK()} replaces "K" and "k" with "0" in grade-equivalent test scores in anticipation of converting to numeric.
These sometimes occur in otherwise numeric standardized test scores.
Report the number of such characters replace.

\code{cleanStar()} removes "*" from ordinal values in anticipation of converting to numeric.
These sometimes occur in otherwise numeric standardized test scores.
Report the number of such characters removed.

\code{cleanPlus()} removes "+" from ordinal values at end of range in anticipation of converting to numeric.
These sometimes occur in otherwise numeric standardized test scores.
Report the number of such characters removed.

\code{cleanCruft()} removes residual non-numeric characters from test scores,
often present due to fat-fingered data entry, in anticipation of converting to numeric.
Report the number of such characters removed.
}
\examples{
c <- c("<K.1", "k.8", "2.5", "_4.3", "6.4", "12.9+", "9.2", "10.1", ">12.9")
cleanNumbers(c)

c = c("<1", "1", "2", "3")
cleanLessthan(c)

c = c("1", "1", "2", ">3")
cleanGreaterthan(c)

c = c("<K.1", "K.8", "1.6", "2.1", "3.0", ">12.9")
cleanK(c)

c = c("*K.1", "K.8", "1.6", "2.1", "3.0", ">12.9")
cleanStar(c)

c = c("*K.1", "K.8", "1.6", "2.1", "3.0", "12.9+")
cleanPlus(c)

c = c("0.1", "_0.8", "1.6", "2.1`", "+3. 0", "12.9")
cleanCruft(c)

}
\seealso{
\code{\link{readr::parse_number()}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
