% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compCorr.R
\name{compCorr}
\alias{compCorr}
\title{Fisher's Z test for comparing two correllation coefficients.}
\usage{
compCorr(n1, r1, n2, r2)
}
\arguments{
\item{n1}{n for correlation 1.}

\item{r1}{r for correlation 1.}

\item{n2}{n for correlation 2.}

\item{r2}{r for correlation 2.}
}
\value{
list containing diff(Z1, Z2) and pval for difference.
}
\description{
None yet.
}
\details{
Test whether two correlation coefficients are statistically different using the difference
between Fisher's Z transforms.

Method cribbed from:
\enumerate{
    \item \url{http://www.fon.hum.uva.nl/Service/Statistics/Two_Correlations.html}
    \item \url{http://ftp.sas.com/techsup/download/stat/compcorr.html}
}
Implemented in R by me.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}

