% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fact2char}
\alias{fact2char}
\title{Coerce Factor Vector to Character Preserving Labels.}
\usage{
fact2char(f)
}
\arguments{
\item{f}{A vector of class factor.}
}
\value{
A vector of class "character".
}
\description{
Coerce a vector of class "factor" to character.
}
\details{
Coerces a vector of class factor to numeric. This function preserves level labels associated
with each element of f.

cf. > unclass(f)
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{fact2num}}
\code{\link{as.character}}
}

