% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{fact2num}
\alias{fact2num}
\title{Coerce Factor Vector to Numeric Preserving Levels.}
\usage{
fact2num(f)
}
\arguments{
\item{f}{A vector of class "factor".}
}
\value{
A vector of class "numeric".
}
\description{
Coerce a vector of class "factor" to numeric.
}
\details{
Coerces a vector of class factor to numeric. This function preserves underlying integer
associated with each element of f.

cf. > unclass(f)
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link{fact2char}}
\code{\link{as.numeric}}
}

