% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDButils-package.R
\docType{package}
\name{FDButils}
\alias{FDButils}
\alias{FDButils-package}
\title{Utility and Convenience Functions.}
\description{
This package is a repository for low-level
    utility functions used by FDB. As a rule,
    no function in this package will
    depend on anything beyond the R 'base' and 'recommended'
    packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davebraze/FDButils}
  \item Report bugs at \url{https://github.com/davebraze/FDButils/issues}
}

}
\author{
\strong{Maintainer}: David Braze \email{davebraze@gmail.com}

}
\keyword{internal}
