% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Greatest common divisor.}
\usage{
gcd(v)
}
\arguments{
\item{v}{A vector of whole numbers}
}
\value{
Returns a numeric value corresponding to the greatest common divisor of the numbers in v.
}
\description{
\code{gcd()} takes a vector of whole numbers and returns their greatest common divisor as a numeric vector of length 1.
}
\details{
\code{gcd()} takes a vector, v, of whole numbers as argument and returns their greatest common
    divisor. Throws an error if any number in v is not a whole number (to within \code{.Machine$double.eps^0.5}).
}
\examples{
gcd(c(48, 96, 144))
}
