% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invDnorm.R
\name{invDnorm}
\alias{invDnorm}
\alias{inv_dnorm}
\title{Positive Inverse Of The Normal Density Function.}
\usage{
invDnorm(x)

inv_dnorm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector the same length as x.
}
\description{
Normal density function is not fully invertible. This function returns the positive
    inverse.
}
\details{
Cribbed from <http://stackoverflow.com/questions/19589191/the-reverse-inverse-of-the-normal-distribution-function-in-r>

\code{inv_dnorm()} is deprecated.
}
\examples{
x <- seq(-4, 4, .1)
y <- dnorm(x)
plot(x,y)
points(invDnorm(y), y, pch=3, col="blue")
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
