% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invLogit.R
\name{invLogit}
\alias{invLogit}
\title{Logits to probabilities.}
\usage{
invLogit(x)
}
\arguments{
\item{x}{A vector on the logit scale.}
}
\value{
A vector on on the probability scale.
}
\description{
A wrapper around \code{\link[stats]{plogis}}.
}
\details{
Convert logit units to probabilities.
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[stats]{plogis}}
\code{\link[base]{exp}} (for converting logits to odds-ratios)
}

