% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOdd.R
\name{isEven}
\alias{isEven}
\title{Are elements of x even numbers?}
\usage{
isEven(n)
}
\arguments{
\item{n}{A numeric vector.}
}
\value{
A logical vector the same length as n. TRUE if n is an integer and is even, FALSE if n
is an integer and not even, NA otherwise.
}
\description{
Return TRUE if n is an integer evenly divisibly by 2.
}
\details{
Test whether elements of a numeric vector are even. For whole numbers return TRUE or FALSE. For decimals return NA.
}
\examples{
n <- 1:10
isEven(n)

m <- seq(1, 10, .5)
isEven(m)
}
\author{
David Braze \email{davebraze@gmail.com}
}

