% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNumeral.R
\name{isNumeral}
\alias{is.numeral}
\alias{isNumeral}
\title{Can character vector be coerced to numeric?}
\usage{
isNumeral(c)

is.numeral(c)
}
\arguments{
\item{c}{A vector of strings.}
}
\value{
A vector of logicals, NA where \code{is.na(c)}.
}
\description{
None yet.
}
\details{
Checks element-wise as to whether each item in a vector of strings can be coerced to numeric.

Basics cribbed from \url{http://rosettacode.org/wiki/Determine_if_a_string_is_numeric#R}

\code{is.numeral()} is deprecated.
}
\examples{
x <- c('a', 'IX', 'twelve', '0001', '2', '3e1', '-4', '5.5', 'Inf', NA)
data.frame(x,
           isNumeral(x),
           as.numeric(x))
}
\author{
David Braze \email{davebraze@gmail.com}
}

