% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOdd.R
\name{isOdd}
\alias{isOdd}
\title{Are elements of x odd numbers?}
\usage{
isOdd(n)
}
\arguments{
\item{n}{A numeric vector.}
}
\value{
A logical vector the same length as n. TRUE if n is an integer and is even, FALSE if n
is an integer and not even, NA otherwise.
}
\description{
Return TRUE if n is an integer that is not evenly divisible by 2.
}
\details{
Test whether elements of numeric vector are odd. For whole numbers return TRUE or FALSE. For decimals return NA.
}
\author{
David Braze \email{davebraze@gmail.com}
}
