% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWholeNumber.R
\name{isWholeNumber}
\alias{is.wholenumber}
\alias{isWholeNumber}
\title{Can Elements Of x Be Interpreted As Whole Numbers?}
\usage{
isWholeNumber(v, tol = .Machine$double.eps^0.5)

is.wholenumber(v, tol)
}
\arguments{
\item{v}{A numeric vector.}

\item{tol}{What counts as close enough to 0? Defaults to .Machine$double.eps^0.5.}
}
\value{
A vector of Logicals.
}
\description{
Determine whether each element of numerical vector can reasonably be interpreted as a
    whole number.
}
\details{
Takes a numeric vector and returns a Logical vector the same length.

\code{is.wholenumber()} is deprecated.
}
\author{
Cribbed from example in \code{\link[base]{integer}}
}

