% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalDist.R
\name{mahalDistC}
\alias{mahalDistC}
\title{Get squared Mahalanobis distances to centroid.}
\usage{
mahalDistC(m, scale = TRUE, use = "complete.obs", center = "mean", ...)
}
\arguments{
\item{m}{A data.frame or matrix. Observations in rows.}

\item{scale}{If TRUE, scale variables before getting mahalanobis
distances.}

\item{use}{Observations to use in computing covariance matrix. Gets
passed to \code{\link[stats]{cov}}.}

\item{center}{Type of univariate center for each variable in \code{m};
"mean" or "median"}

\item{...}{additional arguments passed to
\code{\link[stats]{mahalanobis}}, and possibly on to
\code{\link[base]{solve}}.}
}
\value{
A list with additional class "mahalDist" containing elements:

    \enumerate{

            \item{
                    D2: A vector of squared Mahalanobis distances for
                    observations (rows) in \code{m}.  Incomplete observations return
                    NA.
            }

            \item{
                    vars: A character vector with the column names from
                    \code{m}.
            }

            \item{
                    dim: The number of columns of \code{m}.
            }
    }
}
\description{
For N points, in M dimensions, get squared Mahalanobis
    distances to center.
}
\details{
For each of N points in M dimensions, get squared Mahalanobis
    distances to distribution centroid. This is useful for checking for
    outliers in a multivariate distribution. The squared Mahalanobis
    distances to center for an MV normal distribution with N dimensions
    will follow the chi square distribution with DF equal to N.

    This function is a convenience wrapper around
    \code{\link[stats]{mahalanobis}}, which see. Variables are
    optionally scaled before distances are computed. Incomplete
    observations will return NA.
}
\examples{
m <- matrix(rnorm(400, m=.8, s=.05), nrow=100)
md <- mahalDistC(m)
md$D2
hist(md$D2)
rug(md$D2)

}
\seealso{
\code{\link[stats]{mahalanobis}}

\code{\link[stats]{cov}}

\code{\link[base]{solve}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
