% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalDist.R
\name{mahalDistP}
\alias{mahalDistP}
\title{Get pairwise squared Mahalanobis distances.}
\usage{
mahalDistP(m, covar = NULL, ...)
}
\arguments{
\item{m}{A matrix with observations in rows.}

\item{covar}{The covariance matrix for m.}

\item{...}{Not used.}
}
\value{
An object of class "dist" (see \code{\link[stats]{dist}})
    containing pairwise squared Mahalanobis distances for all
    points in m.
}
\description{
For N points, in M dimensions, get pairwise squared
    Mahalanobis distances between points. Return a square matrix of
    distances.
}
\details{
For each pair of N points in M dimensions, get pairwise
    squared Mahalanobis distances between points. Returns an
    object, \code{d}, of class "dist" (see
    \code{\link[stats]{dist}}) containing pairwise squared
    Mahalanobis distances for all points in m. The method
    attribute of \code{d} is set to "mahalanobis". Obviously,
    \code{d} can get big for large N.

    The full symmetric distance matrix for d can be recovered by
    calling \code{as.matrix(d)}.

    This function is a convenience wrapper around
    \code{\link[stats]{mahalanobis}}, which see.
}
\examples{
m <- matrix(rnorm(200, m=.8, s=.05), nrow=50)
md <- mahalDistP(m)
cluster <- hclust(md)
plot(cluster)

}
\seealso{
\code{\link[stats]{mahalanobis}}

\code{\link[stats]{cov}}

\code{\link[base]{solve}}

\code{\link[stats]{dist}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
