% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.R
\name{nobs}
\alias{nobs}
\title{Number Of Non-missing Values In Vector.}
\usage{
nobs(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
integer, length of x minus the number of NAs in x.
}
\description{
Count non-missing values in v.
}
\details{
Returns the number of non-missing values in vector v. A convenience wrapper around
    \code{sum(!is.na(v))}.
}
\author{
David Braze \email{davebraze@gmail.com}
}
