% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwd.R
\name{openwd}
\alias{openwd}
\title{Open file browser in working directory.}
\usage{
openwd()
}
\value{
TRUE if successful, FALSE otherwise.
}
\description{
Open a file browser in the current working directory, on MS Windows only.
}
\details{
\code{openwd()} opens a standard file browser in the current working directory.
Currently it works on on MS Windows only. Does nothing but issues a
warning on other platforms.
}
\examples{
openwd()
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
