% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{Identify potential outliers in a numeric vector.}
\usage{
outliers(x, method = c("boxplot", "stdev"), coef = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{method}{Method for identifying outliers. Defaults to the
"boxplot" method. Other possibility is "stdev".}

\item{coef}{Multiplier for determining how extreme a value must be
to be flagged as a potential outlier. Defaults to 1.5 for
boxplot method and 2 for stdev method.}
}
\value{
Logical vector the same length as x, indicating for each
    value in x whether it is a possible outlier as determined by
    the specified method and coefficient.
}
\description{
Identify potential outliers in a numeric vector.
}
\details{
Takes a numeric vector, x, and returns a corresponding
    logical vector the same length as x. Values of TRUE in return
    vector indicate corresponding value in x may be an
    outlier. Handy for use with dplyr::mutate().

    If method is boxplot, then values more than
    \code{coef} times the IQR above the 3rd quartile, or more than
    \code{coef} times the IQR below the 1st quartile are flagged
    as possible outliers.

    If method is stdev, then values more than \code{coef} standard
    deviations from the mean are flagged as possible outliers.
}
\examples{
set.seed(1234)
x <- rt(50, df=2)
boxplot(x)
x[outliers(x)]
}
\seealso{
\code{\link[stats]{quantile}}
}
\author{
David Braze \email{davebraze@gmail.com}
}
