% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleMode.R
\name{sampleMode}
\alias{sampleMode}
\title{Get the sample mode of a vector or matrix.}
\usage{
sampleMode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector or matrix.}

\item{na.rm}{If TRUE, remove NAs from x, if FALSE (default) retain them.}
}
\value{
A single value corresponding to the most common value in x.
}
\description{
Return the most common value in a sample or, for numerix x with no duplicate
    values, return the maximum density estimate.
}
\details{
Returns the mode (most common value) of a vector or matrix or, for numerix x with no
    duplicate values, return the maximum density estimate. For non-numeric x with no duplicates
    return NA. When there is more than 1 mode, return the one whose value occurs first in x.
}
\examples{
x0 <- sample(1:26, 5000, replace=TRUE)
sampleMode(x0)

x1 <- sample(letters, 5000, replace=TRUE)
sampleMode(x1)

x2 <- rnorm(100000, m=9.5)
sampleMode(x2)
}
\author{
David Braze \email{davebraze@gmail.com}
}
