% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciNot.R
\name{sciNot}
\alias{sciNot}
\title{Convert numeric vector to character in scientific notation.}
\usage{
sciNot(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Character vector the same length as x, containing
    representation in scientific notation of each element in x.
}
\description{
Convert numeric vector to character in scientific notation.
}
\details{
Takes a numeric vector, x, and returns a corresponding
    character vector the same length as x. Elements of the return
    vector are representations of corresponding elements of x in
    scientific notation. This is handy for use with 'labels'
    argument to ggplot2:: 'scales' functions:

    scales_x_log10(labels=sciNot)
}
\author{
David Braze \email{davebraze@gmail.com}
}
\seealso{
\code{\link[ggplot2]{scale_continuous}}
}

