% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\alias{seM}
\title{Standard error of the mean.}
\usage{
se(x, na.rm = FALSE)

seM(x, na.rm = FALSE)
}
\arguments{
\item{x}{Object to compute SEMs for. Can be vector, matrix or data.frame.}

\item{na.rm}{Specify how to handle missing values.}
}
\value{
Standard error of the mean for x, or each column of x.
}
\description{
Compute standard error of the mean of x.
}
\details{
Returns the standard error of the mean of x, which can be either a vector, matrix or data.frame.
In the latter two cases, SEM is computed column-wise and a vector of values is returned. When x is
a vector, a single value is returned.
}
\author{
David Braze \email{davebraze@gmail.com}
}

