% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seP.R
\name{seP}
\alias{seP}
\title{Standard Error of a Proportion.}
\usage{
seP(x)
}
\arguments{
\item{x}{A vector containing only 1s and 0s. NAs are silently dropped.}
}
\value{
The standard error of a proportion.
}
\description{
None yet.
}
\details{
Returns the standard error of the proportion of 1s in a vector of 1s and 0s. The vector must
contain only 1s and 0s. Missing values are silently dropped. If x is a matrix or data.frame then
seP is applied column-wise and a vector of values is returned.
}
\author{
David Braze \email{davebraze@gmail.com}
}
