% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{series}
\alias{series}
\title{Find regular sequences of integers in numeric vector.}
\usage{
series(v, step = 1, minseries = 2)
}
\arguments{
\item{v}{A numeric vector containing only whole numbers.}

\item{step}{The step size and direction (distance between adjacent items) for identifying
series. Note that the sign of the step matters. Use positive values for monotonically
increasing series and negative values for monotonically decreasing series. Defaults to
1. Use 0 for invariant series (runs).}

\item{minseries}{The minimimum series length. Defaults to 2.}
}
\value{
Numeric matrix with 1 row for each series, and three columns indicating:

\enumerate{
    \item first: Integer at beginning of series.
    \item runstart: Index into \code{v} for start of series.
    \item runlength: Length of series.
}
}
\description{
Find regular sequences of integers in numeric vector v.
}
\details{
Takes a vector of whole numbers and returns a matrix with 3 columns: (1) integer at
beginning of series, (2) index pointing to beginning of series, (3) length of series
}
\author{
David Braze \email{davebraze@gmail.com}
}

