% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp}
\alias{timestamp}
\title{Get a human readable timestamp as string.}
\usage{
timestamp()
}
\value{
A string corresponding to a human readable timestamp: 'YYYYMMDD-HHMMSS'.
}
\description{
Return a human readable timestamp as a string: 'YYYYMMDD-HHMMSS'.
}
\details{
Convenience function to get a human readable timestamp string: 'YYYYMMDD-HHMMSS'.
This is mostly useful for including in filenames.
}
\examples{
timestamp()
}
\author{
David Braze \email{davebraze@gmail.com}
}
