% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanNumbers.R
\name{cleanNumbers}
\alias{cleanNumbers}
\title{Strip all non-numeric characters from string formated test scores.}
\usage{
cleanNumbers(c)
}
\arguments{
\item{c}{A character vector of potentially numeric values.}
}
\value{
A character vector
}
\description{
Remove non-numeric characters from otherwise numeric standardized
test scores.
}
\details{
Remove or replace non-numeric characters that sometimes occur in otherwise numeric standardized
test scores. Several characters are handled specially. These four characters "<>+*"
are simply removed. The letter "K" is \emph{\strong{replaced}} with "0" (zero). All other non-numeric
characters are removed. A count of each type of character removed or replaced is
printed to the console.

\code{cleanNumbers()} calls helper functions \code{\link{scrubc()}} and \code{\link{swapc()}}
to do its work.

Advantages over \code{\link{readr::parse_number()}} are that it provides
explicit alerts as to what characters are being removed or replaced and
in what quantity. Useful in cases where you are working with unfamiliar
data (e.g., client data, or found data).
}
\examples{
c <- c("<K.1", "k.8", "2.5", "_4.3", "6.4", "12.9+", "9.2", "10.1", ">12.9")
cleanNumbers(c)

}
\seealso{
\code{\link{readr::parse_number()}}
\code{\link{scrubc()}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
