% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDButils.R
\docType{package}
\name{FDButils}
\alias{FDButils}
\alias{FDButils-package}
\title{Assorted Utility and Convenience Functions.}
\description{
This package is a repository for stable utility functions used in the FDB lab at
    Haskins.
}

