% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDButils.R
\docType{package}
\name{FDButils}
\alias{FDButils}
\title{Utility and Convenience Functions.}
\description{
This package is a repository for low-level
    utility functions used in the FDB lab. As a rule,
    no function in this package will
    depend on anything beyond the R 'base' and 'recommended'
    packages.
}
