% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDButils-package.R
\docType{package}
\name{FDButils}
\alias{FDButils}
\alias{FDButils-package}
\title{Utility and Convenience Functions.}
\description{
The FDButils package is a repository for
    miscellaneous utility functions used by FDB. It includes,
    few depencies beyond the R 'base' and 'recommended'
    packages, but do see DESCRIPTION file for a handful
    of others. In general, FDButils and its dependencies
    should be pretty stable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davebraze/FDButils}
  \item Report bugs at \url{https://github.com/davebraze/FDButils/issues}
}

}
\author{
\strong{Maintainer}: David Braze \email{davebraze@gmail.com}

}
\keyword{internal}
