% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd0}
\alias{gcd0}
\alias{gcd}
\title{Get the greatest common divisor.}
\usage{
gcd0(a, b)

gcd(v)
}
\arguments{
\item{a}{A whole number}

\item{b}{A whole number}

\item{v}{A vector of whole numbers}
}
\value{
gcd0() returns the greatest common divisor of a and b.

gcd() returns the greatest common divisor of the numbers in v.
}
\description{
gcd0() Takes two whole numbers as arguments and returns their greatest common divisor.

gcd() takes a vector of whole numbers and returns their greatest common divisor.
}
\details{
gcd0() Takes two whole numbers as arguments and returns their greatest common divisor. Throws
    an error if either number is not a whole number.

gcd() takes a vector, v, of whole numbers as argument and returns their greatest common
    divisor. Throws an error if any number in v is not a whole number.
}
\examples{
gcd0(36, 48)
gcd0(-36, 48)
gcd(c(48, 96, 144))
}
\author{
David Braze \email{davebraze@gmail.com}

David Braze \email{davebraze@gmail.com}
}
