% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd0}
\alias{gcd}
\alias{gcd0}
\title{Get greatest common divisor of two numbers.}
\usage{
gcd0(a, b)

gcd(v)
}
\arguments{
\item{a}{A whole number}

\item{b}{A whole number}

\item{v}{A vector of whole numbers}
}
\value{
The greatest common divisor of a and b.

The greatest common divisor of the numbers in v.
}
\description{
Takes two whole numbers as arguments and returns their greatest common divisor.

Takes a vector of whole numbers and returns their greatest common divisor.
}
\details{
Takes two whole numbers as arguments and returns their greatest common divisor. Throws
    an error if either number is not a whole number.

Takes a vector, v, of whole numbers as argument and returns their greatest common
    divisor. Throws an error if any number in v is not a whole number.
}
\examples{
gcd0(36, 48)
gcd0(-36, 48)
gcd(c(48, 96, 144))
}
\author{
David Braze \email{davebraze@gmail.com}

David Braze \email{davebraze@gmail.com}
}

