% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latinsq.R
\name{latinList}
\alias{latinList}
\title{Latin squares.}
\usage{
latinList(nSquares = 3, nDims = 4)
}
\arguments{
\item{nSquares}{Number of latin squares to create.}

\item{nDims}{Size of latin squares.}
}
\value{
A list of nSquares latin squares, each nDim x nDim in size.
}
\description{
Get list of different latin squares.
}
\details{
Get list of different latin squares.
}
\examples{
d <- 5
ll <- latinList(3,d)
ll
ll.mat <- Reduce(rbind,ll)
cond <- letters[1:d]
matrix(cond[ll.mat], ncol=d)
}
\seealso{
magic::latin()
}
\author{
David Braze \email{davebraze@gmail.com}
}
