% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{levelOrder}
\alias{levelOrder}
\title{Reorder factor levels.}
\usage{
levelOrder(f, new.order)
}
\arguments{
\item{f}{A factor.}

\item{new.order}{A character vector containing all and only the levels of f, in the desired order.}
}
\value{
A factor.
}
\description{
None yet.
}
\details{
Reorders the levels of factor f as specified in \code{new.order}, which must contain all and
only the existing levels of f (including unused levels), but in the desired order. If \code{new.order}
and \code{levels(f)} do not contain exactly the same elements, then do nothing.
}
\seealso{
\code{\link{factor}}
\code{\link{levels}}
\code{\link[stats]{relevel}}
}
\author{
Dave Braze \email{davebraze@gmail.com}
}
